fix failed test due to dependency requirement
